Support importing xml-defined custom widgets
Currently one Custom Widget can be mapped to one XML namespace, 
multiple should be possible in the future using wildcard syntax. All
properties are treated as string types, i.e. the API of the custom
widget is not yet evaluated. 

The generated import is NOT using the xml extension, so intermediate
file that re-experts the new type is needed. This is by design, as this
intermediate file is supposed to add application code to the UI
elements.